test(cli): drop legacy jest e2e setup and globally add jest setup types
