+++ packages/@expo/cli/e2e/tsconfig.json
@@ -1,0 +1,5 @@
+{
+  "extends": "../tsconfig.json",
+  "include": ["./__tests__", "./playwright", "./utils", "./setup.ts"],
+  "exclude": []
+}
