+++ packages/@expo/cli/jest.config.js
@@ -11,7 +11,7 @@ module.exports = {
       rootDir: path.resolve(__dirname),
       displayName: require('./package').name,
       roots,
-      setupFiles: ['<rootDir>/e2e/setup.ts'],
+      setupFiles: ['<rootDir>/jest.setup.ts'],
       clearMocks: true,
     },
   ],
